tf_crnn/requirement.txt
tensorflow-gpu==1.12.0
opencv-python
numpy
easydict
imgaug